FR-121 [BE] 이동봉사 현황 수정 API 구현 (#107)
* [FR-121] feat: 이동봉사 현황 수정 API 구현

- 이동 완료 시 이동봉사 현황 상태를 update해서 IN_PROGRESS → COMPLETED으로 수정
- 이동봉사 요청자만 이동봉사 현황을 수정할 수 있도록 검증 로작 구현
  - @AuthenticationPrincipal을 통해 로그인한 사용자 id가 요청자 id와 일치하는지 확인
- 이미 이동봉사 완료 상태(COMPLETED)인 경우에 대한 검증 로직 구현

* [FR-121] feat: 이동봉사 현황 삭제 검증 로직 추가 구현

- 요청자와 봉사자만 이동봉사 현황을 삭제할 수 있도록 검증 로작 구현
  - @AuthenticationPrincipal을 통해 로그인한 사용자 id가 요청자 또는 봉사자의 id와 일치하지 않은지 확인

* [FR-121] feat: 이동봉사 현황 생성 검증 로직 추가 구현

- 나의 아이을 등록한 사용자가 요청자인지 검증하는 로직 수정
  - 기존에는 요청자 또는 봉사자인지 확인했다면, 이제는 요청자인지만 검증하도록 수정
- 요청자와 봉사자만 이동봉사 현황을 삭제할 수 있도록 검증 로작 구현
  - @AuthenticationPrincipal을 통해 로그인한 사용자 id가 요청자 또는 봉사자의 id와 일치하지 않은지 확인

+++ be/src/main/java/com/forpets/be/domain/volunteerworkstatus/service/VolunteerWorkStatusService.java
@@ -26,10 +26,10 @@ public class VolunteerWorkStatusService {
     private final MyAnimalRepository myAnimalRepository;
     private final UserRepository userRepository;
 
-    // 이동봉사 현황 생성
+    // 이동봉사 현황 생성 (약속 잡기)
     @Transactional
     public VolunteerWorkStatusResponseDto createServiceStatus(
-        VolunteerWorkStatusRequestDto requestDto) {
+        VolunteerWorkStatusRequestDto requestDto, User user) {
         MyAnimal myAnimal = myAnimalRepository.findById(requestDto.getMyAnimalId())
             .orElseThrow(() -> new IllegalArgumentException("해당 나의 아이를 찾을 수 없습니다."));
 
@@ -38,10 +38,16 @@ public class VolunteerWorkStatusService {
         User volunteer = userRepository.findById(requestDto.getVolunteerId())
             .orElseThrow(() -> new IllegalArgumentException("해당 봉사자(사용자)를 찾을 수 없습니다."));
 
-        // 나의 아이을 등록한 사용자가 요청을 보낸 사용자도 아니고 봉사자도 아닌 경우
-        if (!(myAnimal.getUser().getId() == requestDto.getRequestorId()
-            || myAnimal.getUser().getId() == requestDto.getVolunteerId())) {
-            throw new IllegalArgumentException("나의 아이를 등록한 사용자와 요청자 또는 봉사자의 정보가 일치하지 않습니다.");
+        // 나의 아이을 등록한 사용자가 요청자인지 검증
+        if (myAnimal.getUser().getId() == null || !myAnimal.getUser().getId()
+            .equals(requestDto.getRequestorId())) {
+            throw new IllegalArgumentException("나의 아이를 등록한 사용자와 요청자의 정보가 일치하지 않습니다.");
+        }
+
+        // 이동봉사 현황은 요청자와 봉사자만 생성 가능하도록 검증
+        if (!user.getId().equals(requestDto.getRequestorId()) && !user.getId()
+            .equals(requestDto.getVolunteerId())) {
+            throw new IllegalArgumentException("이동봉사 현황을 생성할 권한이 없습니다.");
         }
 
         if (volunteerWorkStatusRepository.existsByMyAnimalAndRequestorAndVolunteer(myAnimal,
@@ -69,11 +75,11 @@ public class VolunteerWorkStatusService {
             volunteerWorkStatuses = volunteerWorkStatusRepository.findAll();
         } else if (status.equals("in-progress")) {
             // 진행 중인 이동봉사 현황 조회
-            volunteerWorkStatuses = volunteerWorkStatusRepository.findAllByState(
+            volunteerWorkStatuses = volunteerWorkStatusRepository.findAllByStatus(
                 VolunteerStatus.IN_PROGRESS);
         } else if (status.equals("completed")) {
             // 완료된 이동봉사 현황 조회
-            volunteerWorkStatuses = volunteerWorkStatusRepository.findAllByState(
+            volunteerWorkStatuses = volunteerWorkStatusRepository.findAllByStatus(
                 VolunteerStatus.COMPLETED);
         } else {
             throw new IllegalArgumentException("현황에 대한 상태는 all, in-progress, completed만 입력 가능합니다.");
@@ -112,13 +118,44 @@ public class VolunteerWorkStatusService {
         return VolunteerWorkStatusListResponseDto.from(volunteerWorkStatusResponseDtos, total);
     }
 
-    // 이동봉사 현황 삭제
+    // 이동봉사 현황 수정 (이동 완료)
     @Transactional
-    public void deleteVolunteerWorkStatus(Long volunteerWorkStatusId) {
+    public VolunteerWorkStatusResponseDto updateVolunteerWorkStatus(Long volunteerWorkStatusId,
+        User user) {
         VolunteerWorkStatus volunteerWorkStatus = volunteerWorkStatusRepository.findById(
                 volunteerWorkStatusId)
             .orElseThrow(() -> new IllegalArgumentException("해당 이동봉사 현황을 찾을 수 없습니다."));
-        
+
+        // 이동 완료 버튼은 이동봉사 요청자만 상태 수정 가능하도록 검증
+        if (!user.getId().equals(volunteerWorkStatus.getRequestor().getId())) {
+            throw new IllegalArgumentException("요청자만 이동 완료 처리를 할 수 있습니다.");
+        }
+
+        // 이미 이동봉사 완료 상태(COMPLETED)인 경우
+        if (volunteerWorkStatus.getStatus() == VolunteerStatus.COMPLETED) {
+            throw new IllegalArgumentException("이미 완료 처리된 이동봉사 현황입니다.");
+        }
+
+        volunteerWorkStatus.updateStatus(VolunteerStatus.COMPLETED);
+
+        return VolunteerWorkStatusResponseDto.from(volunteerWorkStatus,
+            volunteerWorkStatus.getMyAnimal(), volunteerWorkStatus.getRequestor(),
+            volunteerWorkStatus.getVolunteer());
+    }
+
+    // 이동봉사 현황 삭제 (약속 취소)
+    @Transactional
+    public void deleteVolunteerWorkStatus(Long volunteerWorkStatusId, User user) {
+        VolunteerWorkStatus volunteerWorkStatus = volunteerWorkStatusRepository.findById(
+                volunteerWorkStatusId)
+            .orElseThrow(() -> new IllegalArgumentException("해당 이동봉사 현황을 찾을 수 없습니다."));
+
+        // 이동봉사 현황은 요청자와 봉사자만 상태 수정 가능하도록 검증
+        if (!user.getId().equals(volunteerWorkStatus.getRequestor().getId()) && !user.getId()
+            .equals(volunteerWorkStatus.getVolunteer().getId())) {
+            throw new IllegalArgumentException("이동봉사 현황을 삭제할 권한이 없습니다.");
+        }
+
         volunteerWorkStatusRepository.delete(volunteerWorkStatus);
     }
 }
